Processor::PreferredFormat
There is a "preferred" format that we can determine first, and if it's
found, return early to avoid additional processing.
