+++ lib/psulib_traject.rb
@@ -13,6 +13,7 @@ module PsulibTraject
   require 'psulib_traject/marc_combining_reader'
   require 'psulib_traject/processors/access_facet'
   require 'psulib_traject/processors/format'
+  require 'psulib_traject/processors/preferred_format'
   require 'psulib_traject/processors/media_type'
   require 'psulib_traject/processors/pub_date'
   require 'psulib_traject/solr_manager'
