+++ lib/psulib_traject/processors/format.rb
@@ -17,18 +17,15 @@ module PsulibTraject::Processors
 
     # Check 949t, leader6 and 007 formats
     def resolve_formats
-      # Check 949t before other formats to avoid overlapping formats,
-      # eg. prefer Juvenile Book vs Book, Statute vs Government Document
-      formats = local_formats
+      formats = field949t
+      preferred_format = PreferredFormat.call(record: record, local_formats: formats)
+      return preferred_format if preferred_format.any?
 
       overlaps = avoid_overlaps
       formats = overlaps if overlaps && formats.empty?
 
       formats = type_of_record if formats.empty?
       formats = physical_description if formats.nil?
-
-      overrides = resolve_overrides
-      formats = overrides if overrides
 
       # If no other values are present, use the default value "Other"
       formats = 'Other' if formats.empty?
@@ -105,7 +102,7 @@ module PsulibTraject::Processors
     end
 
     # Check 949t formats, a record may have multiple 949s with different 949ts
-    def local_formats
+    def field949t
       [].tap do |formats|
         Traject::MarcExtractor.cached('949t').collect_matching_lines(record) do |field, spec, extractor|
           field_949t = extractor.collect_subfields(field, spec).first
@@ -113,16 +110,6 @@ module PsulibTraject::Processors
           formats << format if format
         end
       end
-    end
-
-    # Check other possible formats and prefer over 949t, leader6 and 007 formats
-    def resolve_overrides
-      format = 'Thesis/Dissertation' if thesis?
-      format = 'Newspaper' if newspaper?
-      format = 'Games/Toys' if games?
-      format = 'Proceeding/Congress' if proceeding? || congress?
-      format = 'Book' if book?
-      format
     end
 
     private
@@ -140,44 +127,9 @@ module PsulibTraject::Processors
         field_260b_264b.grep(/\buniversity\b/i).any?
       end
 
-      # Check if newspaper using leader byte 7 and 008
-      def newspaper?
-        record.leader[7] == 's' && record['008'] && record['008'].value[21] == 'n'
-      end
-
-      # Checks if it has a 502, if it does it's considered a thesis
-      def thesis?
-        !record.find { |a| a.tag == '502' }.nil?
-      end
-
-      # Check leader byte 12 and 008 byte 29 for proceeding/congress
-      def proceeding?
-        record.leader[12] == '1' || (record['008'] && record['008'].value[29] == '1')
-      end
-
-      # Checks all $6xx for a $v "congress"
-      def congress?
-        !record.find do |field|
-          field.tag.slice(0) == '6' && field.subfields.find { |sf| sf.code == 'v' && PsulibTraject.regex_to_extract_data_from_a_string(sf.value, /Congress/i) }
-        end.nil?
-      end
-
-      # Checks leader byte 6 and 16, 006 and 008 for games/toys
-      def games?
-        %w[r m].include?(record.leader[6]) &&
-          (%w[g w].include?(record.leader[16]) ||
-           record['006'] && record['006'].value[9] == 'g' ||
-           record['008'] && (%w[g w].include?(record['008'].value[33]) || record['008'].value[26] == 'g'))
-      end
-
       # Checks 006 and 008 for instructional material
       def instructional_material?
         record['006'] && record['006'].value[16] == 'q' || record['008'] && record['008'].value[33] == 'q'
-      end
-
-      # Override for Book when leader(6-7) is 'am' - issue#172
-      def book?
-        record.leader[6] == 'a' && record.leader[7] == 'm' && local_formats.include?('Archives/Manuscripts')
       end
   end
 end
